requirements.txt
numpy==1.24.4
scipy==1.10.1
roboticstoolbox-python[collision]
pythonfmu